Refactored CRUD domain entities

+++ crates/repository/queries/bookmarks/delete_by_id.sql
@@ -1,3 +1,4 @@
 DELETE FROM bookmarks
 WHERE
   id = $1
+  AND user_id = $2
