+++ crates/repository/queries/subscriptions/delete_by_id.sql
@@ -1,3 +1,4 @@
 DELETE FROM subscriptions
 WHERE
   id = $1
+  AND user_id = $2
